requirements.txt
streamlit
Pillow
transformers
torch
pandas
altair